refactor: directory명 변경사항 적용 #26
