+++ monew/src/main/java/com/codeit/team2/monew/module/domain/comment/entity/CommentLike.java
@@ -1,7 +1,7 @@
 package com.codeit.team2.monew.module.domain.comment.entity;
 
 import com.codeit.team2.monew.module.domain.BaseEntity;
-import com.codeit.team2.monew.module.domain.member.entity.User;
+import com.codeit.team2.monew.module.domain.user.entity.User;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.JoinColumn;
