+++ monew/src/main/java/com/codeit/team2/monew/module/domain/notification/service/NotificationService.java
@@ -2,8 +2,8 @@ package com.codeit.team2.monew.module.domain.notification.service;
 
 import com.codeit.team2.monew.module.domain.article.entity.Article;
 import com.codeit.team2.monew.module.domain.comment.entity.Comment;
-import com.codeit.team2.monew.module.domain.member.entity.User;
 import com.codeit.team2.monew.module.domain.notification.entity.Notification;
+import com.codeit.team2.monew.module.domain.user.entity.User;
 import java.util.List;
 
 public interface NotificationService {
