+++ monew/src/main/java/com/codeit/team2/monew/module/domain/subscription/entity/Subscription.java
@@ -2,7 +2,7 @@ package com.codeit.team2.monew.module.domain.subscription.entity;
 
 import com.codeit.team2.monew.module.domain.BaseEntity;
 import com.codeit.team2.monew.module.domain.interest.entity.Interest;
-import com.codeit.team2.monew.module.domain.member.entity.User;
+import com.codeit.team2.monew.module.domain.user.entity.User;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.JoinColumn;
@@ -13,7 +13,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name="subscriptions")
+@Table(name = "subscriptions")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
 public class Subscription extends BaseEntity {
